app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.0.0
numpy==1.25.2
pydantic==2.3.0